refactor: abstract session into inbound and outbound
and delete some useless ssl code
