+++ src/core/session_outbound.c
@@ -1,0 +1,751 @@
+#include "core/session.h"
+#include "core/codec.h"
+
+#include <event2/buffer.h>
+#include <event2/bufferevent.h>
+
+/*
+ * bypass handlers
+ */
+static void on_bypass_remote_event(struct bufferevent *bev, short events,
+				   void *ctx);
+static void on_bypass_remote_read(struct bufferevent *bev, void *ctx);
+
+/*
+ * trojan session handlers
+ */
+static void on_trojan_remote_event(struct bufferevent *bev, short events,
+				   void *ctx);
+static void on_trojan_remote_read(struct bufferevent *bev, void *ctx);
+
+/*
+ * v2ray session handlers
+ */
+static void on_v2ray_remote_event(struct bufferevent *bev, short events,
+				  void *ctx);
+static void on_v2ray_remote_read(struct bufferevent *bev, void *ctx);
+
+void socks5_dest_addr_parse(const uint8_t *cmd, uint64_t cmd_len,
+			    pgs_acl_t *acl, bool *proxy, char **dest_ptr,
+			    int *port)
+{
+	int atype = cmd[3];
+	int offset = 4;
+	char *dest = NULL;
+
+	switch (atype) {
+	case SOCKS5_CMD_IPV4: {
+		assert(cmd_len > 8);
+		dest = (char *)malloc(sizeof(char) * 32);
+		sprintf(dest, "%d.%d.%d.%d", cmd[offset], cmd[offset + 1],
+			cmd[offset + 2], cmd[offset + 3]);
+		offset += 4;
+		break;
+	}
+	case SOCKS5_CMD_HOSTNAME: {
+		offset = 5;
+		int len = cmd[4];
+		assert(cmd_len > len + 4);
+		dest = (char *)malloc(sizeof(char) * (len + 1));
+		memcpy(dest, cmd + 5, len);
+		dest[len] = '\0';
+		offset += len;
+		break;
+	}
+	case SOCKS5_CMD_IPV6: {
+		assert(cmd_len > 20);
+		dest = (char *)malloc(sizeof(char) * 32);
+		sprintf(dest, "%x:%x:%x:%x:%x:%x:%x:%x",
+			cmd[offset] << 8 | cmd[offset + 1],
+			cmd[offset + 2] << 8 | cmd[offset + 3],
+			cmd[offset + 4] << 8 | cmd[offset + 5],
+			cmd[offset + 6] << 8 | cmd[offset + 7],
+			cmd[offset + 8] << 8 | cmd[offset + 9],
+			cmd[offset + 10] << 8 | cmd[offset + 11],
+			cmd[offset + 12] << 8 | cmd[offset + 13],
+			cmd[offset + 14] << 8 | cmd[offset + 15]);
+		offset += 16;
+		break;
+	}
+	default:
+		break;
+	}
+	*dest_ptr = dest;
+	*port = (cmd[offset] << 8) | cmd[offset + 1];
+#ifdef WITH_ACL
+	if (acl != NULL) {
+		bool match = pgs_acl_match_host(acl, dest);
+		if (!match && atype == SOCKS5_CMD_HOSTNAME) {
+			// TODO: reolve the DNS and match the ip again
+		}
+		if (pgs_acl_get_mode(acl) == PROXY_ALL_BYPASS_LIST) {
+			*proxy = !match;
+		} else if (pgs_acl_get_mode(acl) == BYPASS_ALL_PROXY_LIST) {
+			*proxy = match;
+		}
+	}
+#endif
+}
+
+// trojan session context
+pgs_outbound_ctx_trojan_t *
+pgs_outbound_ctx_trojan_new(const uint8_t *encodepass, uint64_t passlen,
+			    const uint8_t *cmd, uint64_t cmdlen)
+{
+	if (passlen != SHA224_LEN * 2 || cmdlen < 3)
+		return NULL;
+	pgs_outbound_ctx_trojan_t *ptr = (pgs_outbound_ctx_trojan_t *)malloc(
+		sizeof(pgs_outbound_ctx_trojan_t));
+	ptr->head_len = passlen + 2 + 1 + cmdlen - 3 + 2;
+	ptr->head = (char *)malloc(sizeof(char) * ptr->head_len);
+
+	// sha224(password) + "\r\n" + cmd[1] + cmd.substr(3) + "\r\n"
+	memcpy(ptr->head, encodepass, passlen);
+	ptr->head[passlen] = '\r';
+	ptr->head[passlen + 1] = '\n';
+	ptr->head[passlen + 2] = cmd[1];
+	memcpy(ptr->head + passlen + 3, cmd + 3, cmdlen - 3);
+	ptr->head[ptr->head_len - 2] = '\r';
+	ptr->head[ptr->head_len - 1] = '\n';
+	return ptr;
+}
+
+void pgs_outbound_ctx_trojan_free(pgs_outbound_ctx_trojan_t *ctx)
+{
+	if (ctx->head)
+		free(ctx->head);
+	ctx->head = NULL;
+	free(ctx);
+	ctx = NULL;
+}
+
+// vmess context
+pgs_outbound_ctx_v2ray_t *pgs_outbound_ctx_v2ray_new(const uint8_t *cmd,
+						     uint64_t cmdlen,
+						     pgs_v2ray_secure_t secure)
+{
+	pgs_outbound_ctx_v2ray_t *ptr = (pgs_outbound_ctx_v2ray_t *)malloc(
+		sizeof(pgs_outbound_ctx_v2ray_t));
+
+	memzero(ptr->local_rbuf, BUFSIZE_16K);
+	memzero(ptr->local_wbuf, BUFSIZE_16K);
+	memzero(ptr->remote_rbuf, BUFSIZE_16K);
+	memzero(ptr->remote_wbuf, BUFSIZE_16K);
+	ptr->cmd = cmd;
+	ptr->cmdlen = cmdlen;
+	ptr->header_sent = false;
+	ptr->header_recved = false;
+	ptr->resp_len = 0;
+	ptr->target_addr_len = 0;
+	ptr->remote_rbuf_pos = 0;
+	ptr->resp_hash = 0;
+	ptr->encryptor = NULL;
+	ptr->decryptor = NULL;
+	ptr->secure = secure;
+
+	return ptr;
+}
+
+void pgs_outbound_ctx_v2ray_free(pgs_outbound_ctx_v2ray_t *ptr)
+{
+	if (ptr->encryptor) {
+		pgs_cryptor_free(ptr->secure, ptr->encryptor);
+	}
+	if (ptr->decryptor) {
+		pgs_cryptor_free(ptr->secure, ptr->decryptor);
+	}
+	ptr->encryptor = NULL;
+	ptr->decryptor = NULL;
+	if (ptr)
+		free(ptr);
+	ptr = NULL;
+}
+
+// outbound
+
+void pgs_session_outbound_free(pgs_session_outbound_t *ptr)
+{
+	if (ptr->bev) {
+		bufferevent_free(ptr->bev);
+	}
+	if (ptr->ctx) {
+		if (IS_TROJAN_SERVER(ptr->config->server_type)) {
+			pgs_outbound_ctx_trojan_free(
+				(pgs_outbound_ctx_trojan_t *)ptr->ctx);
+		}
+		if (IS_V2RAY_SERVER(ptr->config->server_type)) {
+			pgs_outbound_ctx_v2ray_free(
+				(pgs_outbound_ctx_v2ray_t *)ptr->ctx);
+		}
+	}
+	if (ptr->dest)
+		free(ptr->dest);
+
+	ptr->bev = NULL;
+	ptr->ctx = NULL;
+	ptr->dest = NULL;
+	free(ptr);
+	ptr = NULL;
+}
+
+bool pgs_session_trojan_outbound_init(pgs_session_outbound_t *ptr,
+				      const pgs_server_config_t *config,
+				      const uint8_t *cmd, uint64_t cmd_len,
+				      struct event_base *base,
+				      on_event_cb *event_cb,
+				      on_read_cb *read_cb, void *cb_ctx)
+{
+	ptr->config = config;
+
+	ptr->ctx =
+		pgs_outbound_ctx_trojan_new(config->password, 56, cmd, cmd_len);
+
+	// sni
+	const char *sni = config->server_address;
+	pgs_config_extra_trojan_t *tconf =
+		(pgs_config_extra_trojan_t *)config->extra;
+	if (tconf->ssl.sni != NULL) {
+		sni = tconf->ssl.sni;
+	}
+	if (pgs_session_outbound_ssl_bev_init(&ptr->bev, base, tconf->ssl_ctx,
+					      sni))
+		goto error;
+
+	assert(event_cb && read_cb && ptr->bev);
+	bufferevent_setcb(ptr->bev, read_cb, NULL, event_cb, cb_ctx);
+
+	return true;
+
+error:
+	return false;
+}
+
+bool pgs_session_v2ray_outbound_init(pgs_session_outbound_t *ptr,
+				     const pgs_server_config_t *config,
+				     const uint8_t *cmd, uint64_t cmd_len,
+				     struct event_base *base,
+				     on_event_cb *event_cb, on_read_cb *read_cb,
+				     void *cb_ctx)
+{
+	pgs_config_extra_v2ray_t *vconf =
+		(pgs_config_extra_v2ray_t *)config->extra;
+
+	ptr->ctx = pgs_outbound_ctx_v2ray_new(cmd, cmd_len, vconf->secure);
+
+	if (vconf->ssl.enabled && vconf->ssl_ctx) {
+		// ssl + vmess
+		const char *sni = config->server_address;
+		if (vconf->ssl.sni != NULL) {
+			sni = vconf->ssl.sni;
+		}
+		if (pgs_session_outbound_ssl_bev_init(&ptr->bev, base,
+						      vconf->ssl_ctx, sni))
+			goto error;
+	} else {
+		// raw vmess
+		ptr->bev = bufferevent_socket_new(
+			base, -1,
+			BEV_OPT_CLOSE_ON_FREE | BEV_OPT_DEFER_CALLBACKS);
+	}
+
+	assert(event_cb && read_cb && ptr->bev);
+	bufferevent_setcb(ptr->bev, read_cb, NULL, event_cb, cb_ctx);
+	return true;
+
+error:
+	return false;
+}
+
+bool pgs_session_bypass_outbound_init(pgs_session_outbound_t *ptr,
+				      struct event_base *base,
+				      on_event_cb *event_cb,
+				      on_read_cb *read_cb, void *cb_ctx)
+{
+	if (event_cb == NULL || read_cb == NULL)
+		goto error;
+	ptr->ctx = NULL;
+	ptr->bev = bufferevent_socket_new(
+		base, -1, BEV_OPT_CLOSE_ON_FREE | BEV_OPT_DEFER_CALLBACKS);
+	bufferevent_setcb(ptr->bev, read_cb, NULL, event_cb, cb_ctx);
+	bufferevent_enable(ptr->bev, EV_READ);
+	ptr->ready = true;
+	ptr->bypass = true;
+
+	return true;
+error:
+	return false;
+}
+
+pgs_session_outbound_t *pgs_session_outbound_new()
+{
+	pgs_session_outbound_t *ptr = malloc(sizeof(pgs_session_outbound_t));
+	ptr->ready = false;
+	ptr->bypass = false;
+	ptr->dest = NULL;
+	ptr->port = 0;
+	ptr->config = NULL;
+	ptr->bev = NULL;
+	ptr->ctx = NULL;
+
+	return ptr;
+}
+
+bool pgs_session_outbound_init(pgs_session_outbound_t *ptr, bool is_udp,
+			       const pgs_config_t *gconfig,
+			       const pgs_server_config_t *config,
+			       const uint8_t *cmd, uint64_t cmd_len,
+			       pgs_local_server_t *local, void *cb_ctx)
+{
+	ptr->config = config;
+
+	bool proxy = true;
+	// CHECK if all zeros for UDP
+	socks5_dest_addr_parse(cmd, cmd_len, local->acl, &proxy, &ptr->dest,
+			       &ptr->port);
+
+	if (ptr->dest == NULL) {
+		pgs_logger_error(local->logger, "socks5_dest_addr_parse");
+		goto error;
+	}
+
+	if (proxy || is_udp) {
+		if (IS_TROJAN_SERVER(config->server_type)) {
+			if (!pgs_session_trojan_outbound_init(
+				    ptr, config, cmd, cmd_len, local->base,
+				    on_trojan_remote_event,
+				    on_trojan_remote_read, cb_ctx)) {
+				pgs_logger_error(
+					local->logger,
+					"Failed to init trojan outbound");
+				goto error;
+			}
+		} else if (IS_V2RAY_SERVER(config->server_type)) {
+			if (!pgs_session_v2ray_outbound_init(
+				    ptr, config, cmd, cmd_len, local->base,
+				    on_v2ray_remote_event, on_v2ray_remote_read,
+				    cb_ctx)) {
+				pgs_logger_error(
+					local->logger,
+					"Failed to init v2ray outbound");
+				goto error;
+			}
+		}
+
+		bufferevent_enable(ptr->bev, EV_READ);
+
+		bufferevent_socket_connect_hostname(ptr->bev, local->dns_base,
+						    AF_INET,
+						    config->server_address,
+						    config->server_port);
+
+		pgs_logger_debug(local->logger, "connect: %s:%d",
+				 config->server_address, config->server_port);
+	}
+	if (!proxy || is_udp) {
+		if (is_udp) {
+			// do nothing
+		} else {
+			pgs_session_bypass_outbound_init(ptr, local->base,
+							 on_bypass_remote_event,
+							 on_bypass_remote_read,
+							 cb_ctx);
+
+			pgs_logger_info(local->logger, "bypass: %s:%d",
+					ptr->dest, ptr->port);
+			bufferevent_socket_connect_hostname(ptr->bev,
+							    local->dns_base,
+							    AF_INET, ptr->dest,
+							    ptr->port);
+		}
+	}
+
+	return true;
+
+error:
+	return false;
+}
+
+// ===================================
+static void on_bypass_remote_event(struct bufferevent *bev, short events,
+				   void *ctx)
+{
+	pgs_session_t *session = (pgs_session_t *)ctx;
+
+	if (events & BEV_EVENT_CONNECTED) {
+		on_bypass_local_read(session->inbound->bev, ctx);
+	}
+
+	if (events & BEV_EVENT_ERROR)
+		pgs_session_error(
+			session,
+			"Error from bufferevent: on_bypass_remote_event");
+	if (events & (BEV_EVENT_EOF | BEV_EVENT_ERROR)) {
+		bufferevent_free(bev);
+
+		pgs_session_free(session);
+	}
+}
+
+static void on_bypass_remote_read(struct bufferevent *bev, void *ctx)
+{
+	pgs_session_t *session = (pgs_session_t *)ctx;
+	pgs_session_debug(session, "remote read triggered");
+	struct evbuffer *input = bufferevent_get_input(bev);
+	uint64_t data_len = evbuffer_get_length(input);
+	unsigned char *data = evbuffer_pullup(input, data_len);
+
+	if (data_len > 0) {
+		struct bufferevent *inbev = session->inbound->bev;
+		struct evbuffer *inboundw = bufferevent_get_output(inbev);
+		evbuffer_add(inboundw, data, data_len);
+		evbuffer_drain(input, data_len);
+	}
+
+	return;
+
+error:
+	pgs_session_free(session);
+}
+
+/*
+ * trojan
+ */
+static void on_trojan_remote_event(struct bufferevent *bev, short events,
+				   void *ctx)
+{
+	pgs_session_t *session = (pgs_session_t *)ctx;
+
+	if (events & BEV_EVENT_CONNECTED) {
+		pgs_session_t *session = (pgs_session_t *)ctx;
+		const pgs_server_config_t *config = session->outbound->config;
+		const pgs_config_extra_trojan_t *trojanconfig = config->extra;
+		if (trojanconfig->websocket.enabled) {
+			// ws conenct
+			pgs_session_debug(session,
+					  "do_trojan_ws_remote_request");
+			pgs_ws_req(
+				bufferevent_get_output(session->outbound->bev),
+				trojanconfig->websocket.hostname,
+				config->server_address, config->server_port,
+				trojanconfig->websocket.path);
+			pgs_session_debug(session,
+					  "do_trojan_ws_remote_request done");
+		} else {
+			// trojan-gfw
+			// should trigger a local read manually
+			pgs_session_debug(session, "trojan-gfw connected");
+			pgs_outbound_ctx_trojan_t *trojan_s_ctx =
+				session->outbound->ctx;
+			session->outbound->ready = true;
+			// manually trigger a read local event
+			if (session->inbound->state == INBOUND_PROXY) {
+				// TCP
+				on_trojan_local_read(session->inbound->bev,
+						     ctx);
+			} else if (session->inbound->state ==
+					   INBOUND_UDP_RELAY &&
+				   session->inbound->udp_fd != 0) {
+				if (session->inbound->rbuf_pos > 0) {
+					on_udp_read_trojan(
+						session->inbound->udp_rbuf,
+						session->inbound->rbuf_pos,
+						session);
+					session->inbound->rbuf_pos = 0;
+				}
+			}
+		}
+	}
+	if (events & BEV_EVENT_ERROR)
+		pgs_session_error(
+			session,
+			"Error from bufferevent: on_trojan_remote_event");
+	if (events & (BEV_EVENT_EOF | BEV_EVENT_ERROR)) {
+		bufferevent_free(bev);
+
+		pgs_session_free(session);
+	}
+}
+
+/*
+ * outound read handler
+ * it will handle websocket upgrade or 
+ * remote -> decode(ws frame) -> local
+ */
+static void on_trojan_remote_read(struct bufferevent *bev, void *ctx)
+{
+	pgs_session_t *session = (pgs_session_t *)ctx;
+	pgs_session_debug(session, "remote read triggered");
+	struct evbuffer *output = bufferevent_get_output(bev);
+	struct evbuffer *input = bufferevent_get_input(bev);
+
+	uint64_t data_len = evbuffer_get_length(input);
+	unsigned char *data = evbuffer_pullup(input, data_len);
+
+	const pgs_server_config_t *config = session->outbound->config;
+	if (config == NULL) {
+		pgs_session_error(session, "current server config not found");
+		goto error;
+	}
+	pgs_config_extra_trojan_t *trojanconf =
+		(pgs_config_extra_trojan_t *)config->extra;
+	if (!trojanconf->websocket.enabled) {
+		// trojan-gfw
+		trojan_write_local(session, data, data_len);
+		evbuffer_drain(input, data_len);
+		return;
+	}
+	// trojan ws
+	pgs_outbound_ctx_trojan_t *trojan_s_ctx = session->outbound->ctx;
+	if (!session->outbound->ready) {
+		if (!strstr((const char *)data, "\r\n\r\n"))
+			return;
+
+		if (pgs_ws_upgrade_check((const char *)data)) {
+			pgs_session_error(session, "websocket upgrade fail!");
+			on_trojan_remote_event(bev, BEV_EVENT_ERROR, ctx);
+		} else {
+			//drain
+			evbuffer_drain(input, data_len);
+			session->outbound->ready = true;
+			// local buffer should have data already
+			// manually trigger a read local event
+			if (session->inbound->state == INBOUND_PROXY) {
+				on_trojan_local_read(bev, ctx);
+			} else if (session->inbound->state ==
+					   INBOUND_UDP_RELAY &&
+				   session->inbound->udp_fd != 0) {
+				if (session->inbound->rbuf_pos > 0) {
+					on_udp_read_trojan(
+						session->inbound->udp_rbuf,
+						session->inbound->rbuf_pos,
+						session);
+					session->inbound->rbuf_pos = 0;
+				}
+			}
+		}
+	} else {
+		// upgraded, decode it and write to local
+		// read from remote
+		pgs_session_debug(session, "remote -> decode -> local");
+
+		if (data_len < 2)
+			return; // wait next read
+
+		while (data_len > 2) {
+			pgs_ws_resp_t ws_meta;
+			if (pgs_ws_parse_head(data, data_len, &ws_meta)) {
+				// ignore opcode here
+				if (ws_meta.opcode == 0x01) {
+					// write to local
+					trojan_write_local(
+						session,
+						data + ws_meta.header_len,
+						ws_meta.payload_len);
+				}
+
+				if (!ws_meta.fin)
+					pgs_session_debug(
+						session,
+						"frame to be continued..");
+
+				evbuffer_drain(input,
+					       ws_meta.header_len +
+						       ws_meta.payload_len);
+
+				on_session_metrics_recv(
+					session, ws_meta.header_len +
+							 ws_meta.payload_len);
+
+				data_len -= (ws_meta.header_len +
+					     ws_meta.payload_len);
+				data += (ws_meta.header_len +
+					 ws_meta.payload_len);
+			} else {
+				pgs_session_warn(
+					session,
+					"Failed to parse ws header, wait for more data");
+
+				return;
+			}
+		}
+	}
+	return;
+
+error:
+	pgs_session_free(session);
+}
+
+/*
+ * v2ray wss session handlers
+ */
+static void on_v2ray_remote_event(struct bufferevent *bev, short events,
+				  void *ctx)
+{
+	pgs_session_t *session = (pgs_session_t *)ctx;
+
+	if (events & BEV_EVENT_CONNECTED) {
+		const pgs_server_config_t *config = session->outbound->config;
+		const pgs_config_extra_v2ray_t *vconfig = config->extra;
+		if (vconfig->websocket.enabled) {
+			pgs_session_debug(session,
+					  "do_v2ray_ws_remote_request");
+
+			pgs_ws_req(
+				bufferevent_get_output(session->outbound->bev),
+				vconfig->websocket.hostname,
+				config->server_address, config->server_port,
+				vconfig->websocket.path);
+
+			pgs_session_debug(session,
+					  "do_v2ray_ws_remote_request done");
+		} else {
+			pgs_outbound_ctx_v2ray_t *v2ray_s_ctx =
+				session->outbound->ctx;
+			session->outbound->ready = true;
+			pgs_session_debug(session, "v2ray connected");
+			if (session->inbound->state == INBOUND_PROXY) {
+				// TCP
+				on_v2ray_local_read(session->inbound->bev, ctx);
+			} else if (session->inbound->state ==
+					   INBOUND_UDP_RELAY &&
+				   session->inbound->udp_fd) {
+				if (session->inbound->rbuf_pos > 0) {
+					on_udp_read_v2ray(
+						session->inbound->udp_rbuf,
+						session->inbound->rbuf_pos,
+						session);
+					session->inbound->rbuf_pos = 0;
+				}
+			}
+		}
+	}
+	if (events & BEV_EVENT_ERROR)
+		pgs_session_error(
+			session,
+			"Error from bufferevent: on_v2ray_remote_event");
+	if (events & (BEV_EVENT_EOF | BEV_EVENT_ERROR)) {
+		bufferevent_free(bev);
+
+		pgs_session_free(session);
+	}
+}
+static void on_v2ray_remote_read(struct bufferevent *bev, void *ctx)
+{
+	pgs_session_t *session = (pgs_session_t *)ctx;
+	pgs_session_debug(session, "remote read triggered");
+	const pgs_server_config_t *config = session->outbound->config;
+	const pgs_config_extra_v2ray_t *vconfig = config->extra;
+
+	struct evbuffer *output = bufferevent_get_output(bev);
+	struct evbuffer *input = bufferevent_get_input(bev);
+
+	uint64_t data_len = evbuffer_get_length(input);
+	unsigned char *data = evbuffer_pullup(input, data_len);
+
+	pgs_outbound_ctx_v2ray_t *v2ray_s_ctx = session->outbound->ctx;
+
+	if (!vconfig->websocket.enabled) {
+		struct bufferevent *inbev = session->inbound->bev;
+		struct evbuffer *inboundw = bufferevent_get_output(inbev);
+
+		if (!pgs_vmess_parse(session, data, data_len,
+				     (pgs_session_write_fn)vmess_flush_local)) {
+			pgs_session_error(session,
+					  "failed to decode vmess payload");
+			on_v2ray_remote_event(bev, BEV_EVENT_ERROR, ctx);
+			return;
+		}
+		evbuffer_drain(input, data_len);
+		on_session_metrics_recv(session, data_len);
+		return;
+	}
+	// ws
+	if (!session->outbound->ready) {
+		if (!strstr((const char *)data, "\r\n\r\n"))
+			return;
+
+		if (pgs_ws_upgrade_check((const char *)data)) {
+			pgs_session_error(session, "websocket upgrade fail!");
+			on_v2ray_remote_event(bev, BEV_EVENT_ERROR, ctx);
+		} else {
+			evbuffer_drain(input, data_len);
+			session->outbound->ready = true;
+			if (session->inbound->state == INBOUND_PROXY) {
+				on_v2ray_local_read(bev, ctx);
+			} else if (session->inbound->state ==
+					   INBOUND_UDP_RELAY &&
+				   session->inbound->udp_fd != 0) {
+				// UDP
+				if (session->inbound->rbuf_pos > 0) {
+					on_udp_read_v2ray(
+						session->inbound->udp_rbuf,
+						session->inbound->rbuf_pos,
+						session);
+					session->inbound->rbuf_pos = 0;
+				}
+			}
+		}
+	} else {
+		struct bufferevent *inbev = session->inbound->bev;
+		struct bufferevent *outbev = session->outbound->bev;
+
+		struct evbuffer *outboundr = bufferevent_get_input(outbev);
+		struct evbuffer *inboundw = bufferevent_get_output(inbev);
+
+		while (data_len > 2) {
+			pgs_ws_resp_t ws_meta;
+			if (pgs_ws_parse_head(data, data_len, &ws_meta)) {
+				pgs_session_debug(
+					session,
+					"ws_meta.header_len: %d, ws_meta.payload_len: %d, opcode: %d",
+					ws_meta.header_len, ws_meta.payload_len,
+					ws_meta.opcode);
+				// ignore opcode here
+				if (ws_meta.opcode == 0x02) {
+					// decode vmess protocol
+					if (!pgs_vmess_parse(
+						    session,
+						    data + ws_meta.header_len,
+						    ws_meta.payload_len,
+						    (pgs_session_write_fn)
+							    vmess_flush_local)) {
+						pgs_session_error(
+							session,
+							"failed to decode vmess payload");
+						on_v2ray_remote_event(
+							bev, BEV_EVENT_ERROR,
+							ctx);
+						return;
+					}
+				}
+
+				if (!ws_meta.fin)
+					pgs_session_debug(
+						session,
+						"frame to be continue..");
+
+				evbuffer_drain(outboundr,
+					       ws_meta.header_len +
+						       ws_meta.payload_len);
+
+				on_session_metrics_recv(
+					session, ws_meta.header_len +
+							 ws_meta.payload_len);
+
+				data_len -= (ws_meta.header_len +
+					     ws_meta.payload_len);
+				data += (ws_meta.header_len +
+					 ws_meta.payload_len);
+
+			} else {
+				// error parsing websocket data
+				pgs_session_error(
+					session,
+					"failed to decode websocket payload");
+				on_v2ray_remote_event(bev, BEV_EVENT_ERROR,
+						      ctx);
+				return;
+			}
+		}
+	}
+}
